requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
email-validator==2.1.0
